+++ app/Http/Controllers/ProduitController.php
@@ -6,15 +6,16 @@ use App\Models\Produit;
 use App\Http\Requests\StoreProduitRequest;
 use App\Http\Requests\UpdateProduitRequest;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ProduitController extends Controller
 {
 
-    public function __construct()
-    {
-        $this->middleware('auth:api', ['except' => ['listProduit']]);
-    }
+    // public function __construct()
+    // {
+    //     $this->middleware('auth:api', ['except' => ['listProduit']]);
+    // }
 
     /**
      * Display a listing of the resource.
@@ -22,8 +23,8 @@ class ProduitController extends Controller
     public function listProduit()
     {
         //
-        $produit=Produit::all();
-        return response()->json($produit);
+        $produit=Produit::orderBy('id','desc')->get();
+        return response()->json(["response"=>$produit]);
     }
 
     /**
@@ -37,15 +38,25 @@ class ProduitController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreProduitRequest $request)
+    public function store(Request $request)
     {
         //
-        $request->validated();
+        if (!(auth()->check())) return response()->json(["response"=>"Veuillez vous identifier pour cette action !"]);
+
+        if ( (!isset($request->designation) || empty($request->designation) ) || 
+        (!isset($request->prix) || empty($request->prix)) ||
+        (!isset($request->lienImage) || empty($request->lienImage)) ||
+        (!isset($request->qte) || empty($request->qte)) ||
+        (!isset($request->status) || empty($request->status)) ||
+        (!isset($request->categorie_id) || empty($request->categorie_id))) return response()->json(["response" => "Veuillez remplir tout les champs" ]);
+
+
         $produit= new Produit();
         $produit->designation= $request->designation;
-        $produit->prix= $request->prix;
-        $produit->qte= $request->qte;
-        $produit->categorie_id=$request->categorie_id;
+        $produit->prix= floatval($request->prix);
+        $produit->qte= intval($request->qte);
+        $produit->categorie_id=intval($request->categorie_id);
+        $produit->status=$request->status;
 
         if($request->hasFile('lienImage')) {
             $file_extention = $request->lienImage->getClientOriginalExtension();
@@ -55,8 +66,8 @@ class ProduitController extends Controller
         }else{
             $produit->lienImage = "pas d'image téléchargé";
         }
-        $produit->save();
-        return response()->json(['Ajout effectué avec succès'],403);
+        $produit->update();
+        return response()->json(["response"=>'Ajout effectué avec succès']);
     }
 
     /**
@@ -78,15 +89,22 @@ class ProduitController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateProduitRequest $request, Produit $produit)
+    public function update(Request $request)
     {
-        $request->validated();
-        $produit=Produit::FindOrFail($produit->id);
-        $proVerif=Produit::where('id','<>',$produit->id)
-          ->where('designation',$request->designation)->First();
-        if($proVerif){
-            return response()->json(['Cet produit existe déjà'],409);
+        if (!(auth()->check())) return response()->json(["response"=>"Veuillez vous identifier pour cette action !"]);
 
+        if ( (!isset($request->designation) || empty($request->designation) ) || 
+        (!isset($request->prix) || empty($request->prix)) ||
+        (!isset($request->lienImage) || empty($request->lienImage)) ||
+        (!isset($request->qte) || empty($request->qte)) ||
+        (!isset($request->status) || empty($request->status)) ||
+        (!isset($request->categorie_id) || empty($request->categorie_id))) return response()->json(["response" => "Veuillez remplir tout les champs" ]);
+
+        $produit=Produit::find($request->id);
+        $proVerif=Produit::where('id','<>',$request->id)
+          ->where('designation',$request->designation)->first();
+        if($proVerif){
+            return response()->json(["response"=>'Cet produit existe déjà'],409);
         }
 
         $produit->designation= $request->designation;
@@ -99,14 +117,10 @@ class ProduitController extends Controller
             $uploadedFile = $request->file('lienImage');
             $fileName = 'produit' . time() . '.' . $uploadedFile->getClientOriginalExtension();
             $filePath = 'public/photo/produit';
-
             // Stockez le fichier avec un nom personnalisé
             Storage::putFileAs($filePath, $uploadedFile, $fileName);
-
             // Enregistrez le chemin de la nouvelle image dans la base de données
             $produit->lienImage = $filePath . '/' . $fileName;
-
-
         }
         $produit->update();
         return response()->json(['Modification effectué avec succès'],408);
@@ -118,10 +132,12 @@ class ProduitController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Produit $produit)
+    public function destroy(Request $request)
     {
-        $produit=Produit::FindOrFail($produit->id);
+        $produit=Produit::find($request->id);
+        if(!$produit) return response()->json(["response"=>"Produit not found"]); 
         $produit->delete();
-        return response()->json("Suppression effectuee avec succes");
+        return response()->json(["response"=>"Suppression effectuee avec succes"]);
     }
+    
 }
